remove old frontend

+++ backend/app/models/assets/asset_type_control_link.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, UniqueConstraint, Text
+from sqlalchemy import Column, Integer, ForeignKey, UniqueConstraint, Text, Float
 from app.core.base import Base
 
 class AssetTypeControlLink(Base):
@@ -6,7 +6,7 @@ class AssetTypeControlLink(Base):
     id = Column(Integer, primary_key=True, index=True)
     asset_type_id = Column(Integer, ForeignKey("asset_types.id", ondelete="CASCADE"))
     control_id = Column(Integer, ForeignKey("controls.id", ondelete="CASCADE"))
-    score = Column(Integer, nullable=True, default=0)
+    score = Column(Float, nullable=True, default=0)
     justification = Column(Text, nullable=True)
 
     __table_args__ = (
