+++ backend/app/models/assets/asset_type_threat_link.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, UniqueConstraint, Text
+from sqlalchemy import Column, Integer, ForeignKey, UniqueConstraint, Text, Float
 from app.core.base import Base
 
 class AssetTypeThreatLink(Base):
@@ -6,7 +6,7 @@ class AssetTypeThreatLink(Base):
     id = Column(Integer, primary_key=True, index=True)
     asset_type_id = Column(Integer, ForeignKey("asset_types.id", ondelete="CASCADE"))
     threat_id = Column(Integer, ForeignKey("threats.id", ondelete="CASCADE"))
-    score = Column(Integer, nullable=True, default=0)
+    score = Column(Float, nullable=True, default=0)
     justification = Column(Text, nullable=True )
 
     __table_args__ = (UniqueConstraint('asset_type_id', 'threat_id', name='_asset_type_threat_uc'),)
