+++ backend/app/schemas/assets/asset_type_threat_link.py
@@ -21,7 +21,7 @@ class AssetTypeThreatLinkOutDetails(BaseModel):
     id: int
     threat_id: int
     asset_type_id: int
-    score: Optional[int] = None
+    score: Optional[float] = None
     justification: Optional[str] = None
 
     # From Threat table
